add cmd unit test
